requirements.txt
streamlit
markitdown
python-docx
google-genai